+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 ## [Unreleased]
-## [0.1.2] - 2020-10-18
+## [0.1.2] - 2020-10-19
 ### Added
 - Show file name on top of code panels.
 - Use the 'Esc' key to dismiss the code overlay window.
-- [WIP] Use 'Alt+1', 'Alt+2', 'Alt+3' to navigate between File explorer, reading area, and changes list. 
+- Use the 'Esc' key to bring back focus on the timelapse slider.
+- Use 'Alt+1', 'Alt+2', 'Alt+3' to navigate between File explorer, reading area, and changes list.
+- Use the up and down arrow keys to scroll vertically in the overlapping code area.
 
 ### Changed
 - Automatically scroll to the top of the code area after selecting a new file or revision.